move logic to vault filter base class
